backend/requirements.txt
fastapi
uvicorn
langchain-mongodb
langchain-google-genai
langchain-core
langchain
python-dotenv
pymongo
PyMuPDF
langchain-community
python-multipart
aiofiles
sentence-transformers
langchain-huggingface
langchain-openai==0.3.7
langchain-groq==0.2.2
